Remove PowsyblException in LoadOneTransformerTapChanger
