+++ dynawo-simulation/src/main/java/com/powsybl/dynawo/models/loads/LoadOneTransformerTapChanger.java
@@ -7,9 +7,7 @@
  */
 package com.powsybl.dynawo.models.loads;
 
-import com.powsybl.commons.PowsyblException;
 import com.powsybl.dynawo.builders.ModelConfig;
-import com.powsybl.dynawo.models.TransformerSide;
 import com.powsybl.dynawo.models.VarConnection;
 import com.powsybl.dynawo.models.buses.EquipmentConnectionPoint;
 import com.powsybl.dynawo.models.transformers.TapChangerModel;
@@ -22,7 +20,7 @@ import static com.powsybl.dynawo.models.TransformerSide.HIGH_VOLTAGE;
 /**
  * @author Laurent Issertial {@literal <laurent.issertial at rte-france.com>}
  */
-public class LoadOneTransformerTapChanger extends LoadOneTransformer implements TapChangerModel {
+public class LoadOneTransformerTapChanger extends AbstractLoadOneTransformer implements TapChangerModel {
 
     protected LoadOneTransformerTapChanger(Load load, String parameterSetId, ModelConfig modelConfig) {
         super(load, parameterSetId, modelConfig);
@@ -35,11 +33,6 @@ public class LoadOneTransformerTapChanger extends LoadOneTransformer implements 
                 .map(switchOff -> new VarConnection("tapChanger_switchOffSignal1", switchOff))
                 .ifPresent(varConnections::add);
         return varConnections;
-    }
-
-    @Override
-    public List<VarConnection> getTapChangerVarConnections(TransformerSide side) {
-        throw new PowsyblException("LoadOneTransformerTapChanger already have a tap changer");
     }
 
     @Override
